Fix un namespaced concerns.
I have no idea how this worked before, but it only came up as an error
when loading the environment in a way outside of `bundle exec rails c`
and `bundle exec rails s`.

These should be namespaced properly because the models that include
them, expect them under the solidus braintree namespace
