--- app/models/concerns/add_name_validation_concern.rb
@@ -1,6 +1,0 @@
-module AddNameValidationConcern
-  extend ActiveSupport::Concern
-  included do
-    validates :name, presence: true, on: :create
-  end
-end
