--- app/models/concerns/payment_braintree_nonce_concern.rb
@@ -1,6 +1,0 @@
-module PaymentBraintreeNonceConcern
-  extend ActiveSupport::Concern
-  included do
-    attr_accessor :payment_method_nonce
-  end
-end
