--- app/models/concerns/permitted_attributes_concern.rb
@@ -1,5 +1,0 @@
-module PermittedAttributesConcern
-  def payment_attributes
-    super | [:payment_method_nonce]
-  end
-end
