+++ app/models/concerns/solidus_braintree/add_name_validation_concern.rb
@@ -1,0 +1,8 @@
+module SolidusBraintree
+  module AddNameValidationConcern
+    extend ActiveSupport::Concern
+    included do
+      validates :name, presence: true, on: :create
+    end
+  end
+end
