+++ app/models/concerns/solidus_braintree/payment_braintree_nonce_concern.rb
@@ -1,0 +1,8 @@
+module SolidusBraintree
+  module PaymentBraintreeNonceConcern
+    extend ActiveSupport::Concern
+    included do
+      attr_accessor :payment_method_nonce
+    end
+  end
+end
