+++ app/models/concerns/solidus_braintree/permitted_attributes_concern.rb
@@ -1,0 +1,7 @@
+module SolidusBraintree
+  module PermittedAttributesConcern
+    def payment_attributes
+      super | [:payment_method_nonce]
+    end
+  end
+end
