+++ app/models/concerns/solidus_braintree/skip_require_card_numbers_concern.rb
@@ -1,0 +1,14 @@
+module SolidusBraintree
+  module SkipRequireCardNumbersConcern
+    extend ActiveSupport::Concern
+    included do
+      prepend(InstanceMethods)
+    end
+
+    module InstanceMethods
+      def require_card_numbers?
+        super && !self.payment_method.kind_of?(Solidus::Gateway::BraintreeGateway)
+      end
+    end
+  end
+end
